apps/worker/requirements.txt
redis
rq
python-dotenv
requests
openai-whisper
whisperx
boto3